+++ CoffeePeek.Account.Application/Features/Auth/CheckUserExistsByEmail/CheckUserExistsByEmailQuery.cs
@@ -1,7 +1,6 @@
 ﻿using CoffeePeek.Contract.Abstract;
-using CoffeePeek.Contract.Responses;
 using MediatR;
 
 namespace CoffeePeek.Account.Application.Features.Auth.CheckUserExistsByEmail;
 
-public record CheckUserExistsByEmailCommand(string Email) : IRequest<Response<bool>>;
+public record CheckUserExistsByEmailQuery(string Email) : IRequest<Response<bool>>;