+++ CoffeePeek.Account.Application/Features/Auth/RefreshToken/RefreshTokenCommand.cs
@@ -7,5 +7,5 @@ namespace CoffeePeek.Account.Application.Features.Auth.RefreshToken;
 public record RefreshTokenCommand(
     [property: JsonIgnore] Guid UserId,
     string RefreshToken,
-    string DeviceName = "unknown",
-    string IpAddress = "unknown") : IRequest<Response<RefreshTokenResponse>>;
+    [property: JsonIgnore] string DeviceName = "unknown",
+    [property: JsonIgnore] string IpAddress = "unknown") : IRequest<Response<RefreshTokenResponse>>;