+++ CoffeePeek.AccountService/Program.cs
@@ -85,6 +85,7 @@ builder.Services.AddSingleton(MapsterConfiguration.CreateMapper());
 builder.Services.AddScoped<IJWTTokenService, JWTTokenService>();
 builder.Services.AddScoped<IPasswordHasherService, PasswordHasherService>();
 builder.Services.AddScoped<IRoleRepository, RoleRepository>();
+builder.Services.AddScoped<IQueryUserRepository, QueryUserRepository>();
 builder.Services.AddScoped<IExternalAuthService, ExternalAuthService>();
 builder.Services.AddScoped<IAuthService, AuthService>();