+++ CoffeePeek.Contract/Abstract/ErrorResponse.cs
@@ -1,8 +1,7 @@
 ﻿namespace CoffeePeek.Contract.Abstract;
 
-public class ErrorResponse(string errorCode, string message) : Response(success: false, message, null)
+public class ErrorResponse(string message) : Response(success: false, message, null)
 {
-    public string ErrorCode { get; init; } = errorCode;
 #if DEBUG
     public string StackTrace { get; set; }
     public string InnerException { get; set; }